Revert "PK Long->Integer로 변경"
This reverts commit c287d61047f409dca1ea0b9658eb31c7e364bcdf.

+++ src/main/java/kr/ac/bc/comon/springboot/common/domain/FieldEntity.java
@@ -14,7 +14,7 @@ public class FieldEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "FIELD_CD")
-    private Integer fieldCd;
+    private Long fieldCd;
 
     @Column(name = "FIELD_NM", nullable = false)
     private String fieldNm;
